refactor: Agora renderiza direto a pagina cadastro sem verificar o login
