--- src/context/useAuth.jsx
@@ -1,8 +1,0 @@
-import { useContext } from "react";
-import { AuthContext } from "./AuthContext.jsx";
-
-function useAuth() {
-  useContext(AuthContext);
-}
-
-export default useAuth;
